Library for NexusOutputsMgr classes.
Make new ngen_output library for NexusOutputsMgr and subclasses and move
function definitions out of headers and into source files.
