+++ include/utilities/output/PerNexusCsvOutputMgr.hpp
@@ -18,6 +18,9 @@ limitations under the License.
 #ifndef NGEN_PERNEXUSCSVOUTPUTMGR_HPP
 #define NGEN_PERNEXUSCSVOUTPUTMGR_HPP
 
+#include <unordered_map>
+#include <vector>
+#include <fstream>
 #include "NexusOutputsMgr.hpp"
 
 namespace utils
@@ -35,20 +38,12 @@ namespace utils
          * @param nexus_ids Nexus ids for which this instance manages data (in particular, local nexuses when using MPI).
          * @param output_root The output root for written files (as a string).
          */
-        PerNexusCsvOutputMgr(const std::vector<std::string>& nexus_ids, const std::string &output_root) {
-            for(const auto& id : nexus_ids) {
-                nexus_outfiles[id].open(output_root + id + "_output.csv", std::ios::trunc);
-            }
-        }
+        PerNexusCsvOutputMgr(const std::vector<std::string>& nexus_ids, const std::string &output_root);
 
         /**
          * Commit writes by flushing output file `ofstream`s to filesystem.
          */
-        void commit_writes() override {
-            for (auto &f : nexus_outfiles) {
-                f.second.flush();
-            }
-        }
+        void commit_writes() override;
 
         /**
          * Receive a data entry for this nexus, specifying details including the formulation id.
@@ -59,12 +54,7 @@ namespace utils
          * @param flow_data_at_t The nexus flow contribution at this time index (the main data to write).
          */
         void receive_data_entry(const std::string &formulation_id, const std::string &nexus_id,
-                                const time_marker &data_time_marker, const double flow_data_at_t) override {
-            if (formulation_id != get_default_formulation_id()) {
-                throw std::runtime_error("Cannot write data entry for non-default formulation " + formulation_id + " for nexus " + nexus_id + " when per-nexus CSV output is enabled.");
-            }
-            nexus_outfiles.at(nexus_id) << data_time_marker.sim_time_index << ", " << data_time_marker.time_stamp << ", " << flow_data_at_t << '\n';
-        }
+                                const time_marker &data_time_marker, double flow_data_at_t) override;
 
     private:
         std::unordered_map<std::string, std::ofstream> nexus_outfiles;
