+++ test/CMakeLists.txt
@@ -279,7 +279,7 @@ if (NGEN_WITH_MPI)
     # TODO: this might need more
 
     target_include_directories(test_mpi_unit PUBLIC include/utils)
-    target_link_libraries(test_mpi_unit PUBLIC gtest NGen::core NGen::core_nexus NetCDF MPI::MPI_C)
+    target_link_libraries(test_mpi_unit PUBLIC gtest NGen::core NGen::core_nexus NGen::output NetCDF MPI::MPI_C)
     target_compile_definitions(test_mpi_unit PUBLIC NGEN_WITH_MPI NGEN_MPI_UNIT_TESTS)
 
     #set_target_properties(test_mpi_unit PROPERTIES CROSSCOMPILING_EMULATOR '${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 2')
@@ -553,6 +553,7 @@ ngen_add_test(
         NGen::mdframe
         NGen::logging
         NGen::ngen_bmi
+        NGen::output
         testbmicppmodel
 )
